+++ tests/test_app/TestApp/config/recaptcha.php
@@ -6,26 +6,9 @@
  * When you have keys, set them below and RENAME THIS FILE in "recaptcha.php".
  * Remember: keys for forms and keys for emails ARE DIFFERENT.
  */
-
 return ['Recaptcha' => [
-    /**
-     * Form keys
-     * @see https://www.google.com/recaptcha/admin
-     */
-    'Form' => [
-        //Form public key
-        'public' => 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-        //Form private key
-        'private' => 'bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
-    ],
-    /**
-     * Mail keys
-     * @see http://www.google.com/recaptcha/mailhide/apikey
-     */
-    'Mail' => [
-        //Mail public key
-        'public' => 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
-        //Mail private key
-        'private' => 'bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
-    ],
+    //Form public key
+    'public' => 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+    //Form private key
+    'private' => 'bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb',
 ]];